Separate data_table_test util code into its own file.
This yields a cleaner separation between common testing code
and actual tests (a distinction which is important in Google's
internal build system).
